+++ ReleaseNotes/release_notes.md
@@ -1,6 +1,8 @@
 ## New features
 
  1. Improve a little bit the prinntout of `Ostap::Math::Modment_` objects in python
+ 1. Add Cornish-Fisher estimate for approximate quantile functions
+ 1. Tiny fix for `Ostap::Math::probit`
  
 
 ## Backward-incompatible